Tp finalizado parte 6

+++ TP Nº1/libreriaFunciones.h
@@ -50,4 +50,4 @@ int multiplicacionOperandos(int, int);
  * \return int El resultado del factorial
  *
  */
-int factorialOperandos(int);
+long factorialOperandos(int);
